refactor: group distributed runtime modules
